Many comments
The most important thing is to prove that the surrogate model is really good, and does not overestimate the rating of films. Instead of doing experiments with it, you should show for the whole database what's the predicted rating.

Additional problem is that you are reducing so much the number of tropes, what would be the average rating for movies with that number of tropes?

You really have to justify and discuss all decisions, @raiben. That's the most important thing in this paper. And please check all my comments (these and others) and address them.