Inverting the chart
Probably it makes more sense to invert and use number of tropes as x, since that is our independent variable, and this chart conveys much more information. For instance, we can see that if there are 400 tropes, our rating is going to be better than 6, and with 300 tropes you will always "pass". But you are using 30 tropes, which is not such a guarantee of success.

The year makes sense, but only to show that, on average, movies with 30 tropes tend to be old. And again points to not using tropes explicitly as representation for a movie, but something like tropes2vec.

All these things should be discussed on the paper, @raiben